Removed uses of deprecated Date constructor in tests using a builder for dates and made balance a float

+++ src/main/java/bank/InMemoryTransactions.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class InMemoryTransactions implements Transactions {
 
-    public static final int INITIAL_BALANCE = 0;
+    public static final float INITIAL_BALANCE = 0;
     private final List<Transaction> transactions;
     private final SystemDate systemDate;
 
@@ -32,7 +32,7 @@ public class InMemoryTransactions implements Transactions {
 
     private List<StatementLine> getStatementLinesInTransactionsOrder() {
         List<StatementLine> statementLines = new ArrayList<>();
-        int balance = INITIAL_BALANCE;
+        float balance = INITIAL_BALANCE;
         for (final Transaction transaction : transactions) {
             balance += transaction.amount();
             statementLines.add(new StatementLine(transaction, balance));
